Misc compiler errors fixed and formatting issues
compile, core dumps on print_hash call inside memstats, need to debug
print_hash

+++ meminfo_hash.c
@@ -1,11 +1,14 @@
 /* meminfo_hash.c : Implementation file for int hashing
  * By: David Taylor, et. all
  */
-/* Check this hash function */
 
-uint32_t meminfo_hash(void* item){
+#include<stdint.h>
+
+uint32_t meminfo_hash(void* item)
+{
     uint32_t itemval = (uint32_t) item;
-    uint32_t returnval = item * (item +1) % 81181;
+    uint32_t returnval = (uint32_t)item * ((uint32_t)item + 1 ) % 81181;
+
     return returnval;
 }
